improvements: further symbol coverage, /W4 TU compiling highlighting errors, casts hiding types (WIP), general tool improvements

+++ src/GetMemorySize.c
@@ -22,7 +22,7 @@ int GetMemorySize(void)
             total += info.RegionSize;
         address += info.RegionSize;
     }
-    while ( (unsigned int)address < 0x80000000 );
+    while ( (uintptr_t)address < 0x80000000 );
 
     return total;
 }
